python-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==11.0.0
supabase==2.3.1
